+++ internal/timeutil/time.go
@@ -17,13 +17,14 @@ const (
 
 var unix = time.Unix(0, 0)
 
-// UnmarshalInterval up to ±292 years.
-func UnmarshalInterval(n int64) time.Duration {
-	return time.Duration(n)
+// MicrosecondsToDuration returns time.Duration from given microseconds
+func MicrosecondsToDuration(n int64) time.Duration {
+	return time.Duration(n) * time.Microsecond
 }
 
-func MarshalInterval(d time.Duration) int64 {
-	return int64(d)
+// DurationToMicroseconds returns microseconds from given time.Duration
+func DurationToMicroseconds(d time.Duration) int64 {
+	return int64(d / time.Microsecond)
 }
 
 // UnmarshalDate up to 11761191-01-20 00:00:00 +0000 UTC.