+++ table/table_regress_test.go
@@ -7,12 +7,13 @@ import (
 	"context"
 	"os"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/ydb-platform/ydb-go-sdk/v3/table"
-
 	"github.com/ydb-platform/ydb-go-sdk/v3"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 )
 
 type issue229Struct struct{}
@@ -48,4 +49,32 @@ func connect(t *testing.T) ydb.Connection {
 		ydb.WithAccessTokenCredentials(os.Getenv("YDB_ACCESS_TOKEN_CREDENTIALS")))
 	require.NoError(t, err)
 	return db
+}
+
+func TestIssue259IntervalFromDuration(t *testing.T) {
+	// https://github.com/ydb-platform/ydb-go-sdk/issues/259
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	db := connect(t)
+	err := db.Table().DoTx(ctx, func(ctx context.Context, tx table.TransactionActor) error {
+		res, err := tx.Execute(ctx, `DECLARE $ts as Interval;
+			$ten_micro = CAST(10 as Interval);
+			SELECT $ts == $ten_micro, $ten_micro;`, table.NewQueryParameters(
+			table.ValueParam(`$ts`, types.IntervalValueFromDuration(10*time.Microsecond)),
+		))
+		require.NoError(t, err)
+		require.NoError(t, res.NextResultSetErr(ctx))
+		require.True(t, res.NextRow())
+
+		var (
+			val      bool
+			tenMicro time.Duration
+		)
+		require.NoError(t, res.Scan(&val, &tenMicro))
+		require.True(t, val)
+		require.Equal(t, 10*time.Microsecond, tenMicro)
+		return nil
+	}, table.WithTxSettings(table.TxSettings(table.WithSerializableReadWrite())))
+	require.NoError(t, err)
 }